engine/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
celery>=5.3.6
redis>=5.0.1
pydantic>=2.6.2
playwright>=1.41.0
playwright-stealth>=1.0.6
supabase>=2.3.4
python-dotenv>=1.0.1
httpx>=0.26.0
httpx>=0.26.0
google-search-results>=2.4.2
python-amazon-paapi>=3.4.1
